Refactoring and WIP
