+++ rust/agents/relayer/src/msg/serial_submitter.rs
@@ -14,7 +14,9 @@ use tracing::{debug, error, info, info_span, instrument, instrument::Instrumente
 use hyperlane_base::{db::HyperlaneDB, CachingMailbox, CoreMetrics};
 use hyperlane_core::{HyperlaneChain, HyperlaneDomain, Mailbox, U256};
 
-use crate::msg::PendingMessage;
+use crate::msg::{
+    PendingMessage, PendingOperation, PendingOperationState, TxPrepareResult, TxRunResult,
+};
 
 use super::{
     gas_payment::GasPaymentEnforcer, metadata::BaseMetadataBuilder, metadata::MetadataBuilder,
@@ -75,24 +77,15 @@ use super::{
 #[derive(Debug, new)]
 pub(crate) struct SerialSubmitter {
     /// Receiver for new messages to submit.
-    rx: mpsc::UnboundedReceiver<PendingMessage>,
+    rx: mpsc::UnboundedReceiver<Box<dyn PendingOperation>>,
     /// Messages waiting for their turn to be dispatched. The SerialSubmitter
     /// can only dispatch one message at a time, so this queue could grow.
     #[new(default)]
-    run_queue: BinaryHeap<Reverse<PendingMessage>>,
-    /// Mailbox on the destination chain.
-    mailbox: CachingMailbox,
-    /// Used to construct the ISM metadata needed to verify a message.
-    metadata_builder: BaseMetadataBuilder,
-    /// Interface to agent rocks DB for e.g. writing delivery status upon
-    /// completion.
+    run_queue: BinaryHeap<Reverse<Box<dyn PendingOperation>>>,
+    /// Database for the origin domain
     db: HyperlaneDB,
     /// Metrics for serial submitter.
     metrics: SerialSubmitterMetrics,
-    /// Used to determine if messages have made sufficient gas payments.
-    gas_payment_enforcer: Arc<GasPaymentEnforcer>,
-    /// Hard limit on transaction gas when submitting a transaction.
-    transaction_gas_limit: Option<U256>,
 }
 
 impl SerialSubmitter {
@@ -140,226 +133,65 @@ impl SerialSubmitter {
             .run_queue_length_gauge
             .set(self.run_queue.len() as i64);
 
-        // check if the next message is going to be processable
-        if let Some(Reverse(PendingMessage {
-            next_attempt_after: Some(retry_after),
-            ..
-        })) = self.run_queue.peek()
-        {
-            if Instant::now() < *retry_after {
-                return Ok(());
-            }
-        }
-
         // Pick the next message to try processing.
-        let mut msg = match self.run_queue.pop() {
-            Some(m) => m.0,
+        let mut op = match self.run_queue.pop() {
+            Some(op) => op.0,
             None => return Ok(()),
         };
 
-        match self.process_message(&msg).await {
-            Ok(true) => {
-                info!(msg=%msg.message, "Message processed");
-                self.record_message_process_success(&msg)?;
+        // in the future we could pipeline this so that the next operation is being
+        // prepared while the current one is being submitted
+        match op.prepare().await {
+            TxPrepareResult::Ready => {},
+            TxPrepareResult::NotReady => {
+                self.run_queue.push(Reverse(op));
                 return Ok(());
             }
-            Ok(false) => {
-                info!(msg=%msg.message, "Message not processed");
+            TxPrepareResult::Retry => {
+                self.metrics.txs_failed.inc();
+                self.run_queue.push(Reverse(op));
+                return Ok(());
             }
-            // We expect this branch to be hit when there is unexpected behavior -
-            // defined behavior like gas estimation failing will not hit this branch.
-            Err(error) => {
-                error!(msg=%msg.message, ?error, "Error occurred when attempting to process message");
+            TxPrepareResult::Failure => {
+                self.metrics.txs_failed.inc();
+                return Ok(());
+            }
+            TxPrepareResult::CriticalFailure(e) => {
+                self.metrics.txs_failed.inc();
+                return Err(e);
+            }
+        };
+
+        match op.submit().await {
+            TxRunResult::Success => self.metrics.txs_processed.inc(),
+            TxRunResult::Failure => self.metrics.txs_failed.inc(),
+            TxRunResult::Retry => {
+                self.metrics.txs_failed.inc();
+                self.run_queue.push(Reverse(op));
+            }
+            TxRunResult::CriticalFailure(e) => {
+                self.metrics.txs_failed.inc();
+                return Err(e);
             }
         }
-
-        // The message was not processed, so increment the # of retries and add
-        // it back to the run_queue so it will be processed again at some point.
-        msg.num_retries += 1;
-        msg.last_attempted_at = Instant::now();
-        msg.next_attempt_after =
-            Self::calculate_msg_backoff(msg.num_retries).map(|dur| msg.last_attempted_at + dur);
-        self.run_queue.push(Reverse(msg));
 
         Ok(())
     }
 
-    /// Returns the message's status. If the message is processed, either by a
-    /// transaction in this fn or by a view call to the Mailbox contract
-    /// discovering the message has already been processed, Ok(true) is
-    /// returned. If this message is unable to be processed, either due to
-    /// failed gas estimation or an insufficient gas payment, Ok(false) is
-    /// returned.
-    #[instrument(skip(self))]
-    async fn process_message(&self, msg: &PendingMessage) -> Result<bool> {
-        // If the message has already been processed, e.g. due to another relayer having
-        // already processed, then mark it as already-processed, and move on to
-        // the next tick.
-        //
-        // TODO(webbhorn): Make this robust to re-orgs on mailbox.
-        if self.mailbox.delivered(msg.message.id()).await? {
-            debug!("Message already processed");
-            return Ok(true);
-        }
-
-        // The Mailbox's `recipientIsm` function will revert if
-        // the recipient is not a contract. This can pose issues with
-        // our use of the RetryingProvider, which will continuously retry
-        // the eth_call to the `recipientIsm` function.
-        // As a workaround, we avoid the call entirely if the recipient is
-        // not a contract.
-        const CTX: &str = "When fetching ISM";
-        let provider = self.mailbox.provider();
-        if !provider
-            .is_contract(&msg.message.recipient)
-            .await
-            .context(CTX)?
-        {
-            info!(
-                recipient=?msg.message.recipient,
-                "Could not fetch metadata: Recipient is not a contract"
-            );
-            return Ok(false);
-        }
-
-        let ism_address = self
-            .mailbox
-            .recipient_ism(msg.message.recipient)
-            .await
-            .context(CTX)?;
-
-        let Some(metadata) = self.metadata_builder
-            .build(ism_address, &msg.message)
-            .await?
-        else {
-            info!("Could not fetch metadata");
-            return Ok(false)
-        };
-
-        // Estimate transaction costs for the process call. If there are issues, it's
-        // likely that gas estimation has failed because the message is
-        // reverting. This is defined behavior, so we just log the error and
-        // move onto the next tick.
-        let tx_cost_estimate = match self
-            .mailbox
-            .process_estimate_costs(&msg.message, &metadata)
-            .await
-        {
-            Ok(tx_cost_estimate) => tx_cost_estimate,
-            Err(error) => {
-                info!(?error, "Error estimating process costs");
-                return Ok(false);
-            }
-        };
-
-        // If the gas payment requirement hasn't been met, move to the next tick.
-        let Some(gas_limit) = self
-            .gas_payment_enforcer
-            .message_meets_gas_payment_requirement(&msg.message, &tx_cost_estimate)
-            .await?
-        else {
-            info!(?tx_cost_estimate, "Gas payment requirement not met yet");
-            return Ok(false);
-        };
-
-        // Go ahead and attempt processing of message to destination chain.
-        debug!(?gas_limit, "Ready to process message");
-
-        // TODO: consider differentiating types of processing errors, and pushing to the
-        //  front of the run queue for intermittent types of errors that can
-        //  occur even if a message's processing isn't reverting, e.g. timeouts
-        //  or txs being dropped from the mempool. To avoid consistently retrying
-        //  only these messages, the number of retries could be considered.
-
-        let gas_limit = tx_cost_estimate.gas_limit;
-
-        if let Some(max_limit) = self.transaction_gas_limit {
-            if gas_limit > max_limit {
-                info!("Message delivery estimated gas exceeds max gas limit");
-                return Ok(false);
-            }
-        }
-
-        // We use the estimated gas limit from the prior call to
-        // `process_estimate_costs` to avoid a second gas estimation.
-        let outcome = self
-            .mailbox
-            .process(&msg.message, &metadata, Some(gas_limit))
-            .await?;
-
-        // TODO(trevor): Instead of immediately marking as processed, move to a
-        //  verification queue, which will wait for finality and indexing by the
-        //  mailbox indexer and then mark as processed (or eventually retry if
-        //  no confirmation is ever seen).
-
-        self.gas_payment_enforcer
-            .record_tx_outcome(&msg.message, outcome)?;
-        if outcome.executed {
-            info!(
-                hash=?outcome.txid,
-                rq_sz=?self.run_queue.len(),
-                "Message successfully processed by transaction"
-            );
-            Ok(true)
-        } else {
-            info!(
-                hash=?outcome.txid,
-                "Transaction attempting to process transaction reverted"
-            );
-            Ok(false)
-        }
-    }
-
-    /// Record in HyperlaneDB and various metrics that this process has observed
-    /// the successful processing of a message. An `Ok(())` value returned by
-    /// this function is the 'commit' point in a message's lifetime for
-    /// final processing -- after this function has been seen to
-    /// `return Ok(())`, then without a wiped HyperlaneDB, we will never
-    /// re-attempt processing for this message again, even after the relayer
-    /// restarts.
-    fn record_message_process_success(&mut self, msg: &PendingMessage) -> Result<()> {
-        self.db.mark_nonce_as_processed(msg.message.nonce)?;
-        self.metrics.max_submitted_nonce =
-            std::cmp::max(self.metrics.max_submitted_nonce, msg.message.nonce);
-        self.metrics
-            .processed_gauge
-            .set(self.metrics.max_submitted_nonce as i64);
-        self.metrics.messages_processed_count.inc();
+    fn record_message_process_success(&mut self) -> Result<()> {
+        self.metrics.txs_processed.inc();
         Ok(())
-    }
-
-    /// Get duration we should wait before re-attempting to deliver a message
-    /// given the number of retries.
-    fn calculate_msg_backoff(num_retries: u32) -> Option<Duration> {
-        Some(Duration::from_secs(match num_retries {
-            i if i < 1 => return None,
-            // wait 10s for the first few attempts; this prevents thrashing
-            i if (1..12).contains(&i) => 10,
-            // wait 90s to 19.5min with a linear increase
-            i if (12..24).contains(&i) => (i as u64 - 11) * 90,
-            // exponential increase + 30 min; -21 makes it so that at i = 32 it will be
-            // ~60min timeout (64min to be more precise).
-            i => (2u64).pow(i - 21) + 60 * 30,
-        }))
     }
 }
 
 #[derive(Debug)]
 pub(crate) struct SerialSubmitterMetrics {
     run_queue_length_gauge: IntGauge,
-    processed_gauge: IntGauge,
-    messages_processed_count: IntCounter,
-
-    /// Private state used to update actual metrics each tick.
-    max_submitted_nonce: u32,
+    txs_processed: IntCounter,
 }
 
 impl SerialSubmitterMetrics {
-    pub fn new(
-        metrics: &CoreMetrics,
-        origin: &HyperlaneDomain,
-        destination: &HyperlaneDomain,
-    ) -> Self {
+    pub fn new(metrics: &CoreMetrics, destination: &HyperlaneDomain) -> Self {
         let origin = origin.name();
         let destination = destination.name();
         Self {
@@ -368,15 +200,9 @@ impl SerialSubmitterMetrics {
                 destination,
                 "run_queue",
             ]),
-            messages_processed_count: metrics
-                .messages_processed_count()
-                .with_label_values(&[origin, destination]),
-            processed_gauge: metrics.last_known_message_nonce().with_label_values(&[
-                "message_processed",
-                origin,
-                destination,
-            ]),
-            max_submitted_nonce: 0,
+            txs_processed: metrics
+                .transactions_processed()
+                .with_label_values(&[destination]),
         }
     }
 }
